Rename the extern and the external code symbols
